+++ ts/src/lib/action-type.ts
@@ -2,7 +2,8 @@ import { CaseStyle } from "kryo/case-style";
 import { TsEnumType } from "kryo/types/ts-enum";
 
 export enum ActionType {
-  Unknown = 0x00,
+  Unknown = 0x100,
+  Error = 0x101,
   Add = 0x0a,
   Add2 = 0x47,
   And = 0x10,