+++ ts/src/test/raw.spec.ts
@@ -13,6 +13,8 @@ const JSON_READER: JsonReader = new JsonReader();
 const sampleNames: ReadonlyArray<string> = [
   "define-function",
   "trace",
+  "error",
+  "error-with-message",
 ];
 
 describe("$Action.read", function () {